+++ src/Plugin.php
@@ -55,7 +55,7 @@ class Plugin extends BasePlugin
     public bool $hasCpSection = true;
     public bool $hasCpSettings = true;
     public bool $hasReadOnlyCpSettings = true;
-    public string $schemaVersion = '1.4.0';
+    public string $schemaVersion = '1.4.3';
 
     public function init(): void
     {